+++ src/database/MoAI.Database.Shared/Entities/ClassifyEntity.cs
@@ -37,7 +37,7 @@ public partial class ClassifyEntity : IFullAudited
     /// <summary>
     /// 创建人.
     /// </summary>
-    public int CreateUserId { get; set; }
+    public long CreateUserId { get; set; }
 
     /// <summary>
     /// 创建时间.
@@ -47,7 +47,7 @@ public partial class ClassifyEntity : IFullAudited
     /// <summary>
     /// 更新人.
     /// </summary>
-    public int UpdateUserId { get; set; }
+    public long UpdateUserId { get; set; }
 
     /// <summary>
     /// 更新时间.
