+++ src/database/MoAI.Database.Shared/Entities/UserEntity.cs
@@ -72,7 +72,7 @@ public partial class UserEntity : IFullAudited
     /// <summary>
     /// 创建人.
     /// </summary>
-    public int CreateUserId { get; set; }
+    public long CreateUserId { get; set; }
 
     /// <summary>
     /// 创建时间.
@@ -82,7 +82,7 @@ public partial class UserEntity : IFullAudited
     /// <summary>
     /// 最后修改人.
     /// </summary>
-    public int UpdateUserId { get; set; }
+    public long UpdateUserId { get; set; }
 
     /// <summary>
     /// 更新时间.
