+++ tool/PostgresScaffold/Database/Entities/UserOauthConnectionEntity.cs
@@ -17,12 +17,12 @@ public partial class UserOauthConnectionEntity : IFullAudited
     /// <summary>
     /// id.
     /// </summary>
-    public int Id { get; set; }
+    public long Id { get; set; }
 
     /// <summary>
     /// 用户id.
     /// </summary>
-    public int UserId { get; set; }
+    public long UserId { get; set; }
 
     /// <summary>
     /// 供应商id,对应oauth_connection表.
@@ -37,7 +37,7 @@ public partial class UserOauthConnectionEntity : IFullAudited
     /// <summary>
     /// 创建人.
     /// </summary>
-    public int CreateUserId { get; set; }
+    public long CreateUserId { get; set; }
 
     /// <summary>
     /// 创建时间.
@@ -47,7 +47,7 @@ public partial class UserOauthConnectionEntity : IFullAudited
     /// <summary>
     /// 最后修改人.
     /// </summary>
-    public int UpdateUserId { get; set; }
+    public long UpdateUserId { get; set; }
 
     /// <summary>
     /// 更新时间.
